requirements.txt
tweepy==3.10
httpx
SQLAlchemy
mariadb
python-dotenv
pyyaml
